清除无效引入包名

+++ chezuService/src/main/java/com/idata365/service/CalFamilyMonthOrderService.java
@@ -37,7 +37,7 @@ public class CalFamilyMonthOrderService {
 		{
 			taskFamilyOrder.setOrderNo(1);
 		}else {
-			if(preOrder.getScore()==taskFamilyOrder.getScore()) {
+			if(preOrder.getScore().doubleValue()==taskFamilyOrder.getScore().doubleValue()) {
 				taskFamilyOrder.setOrderNo(preOrder.getOrderNo());
 			}else {
 				taskFamilyOrder.setOrderNo(preOrder.getOrderNo()+1);
